Refactor conformance rule suits to minimize code duplication
