+++ conformance/src/test/scala/za/co/absa/enceladus/conformance/interpreter/rules/SparkSessionRuleSuite.scala
@@ -15,52 +15,41 @@
 
 package za.co.absa.enceladus.conformance.interpreter.rules
 
-import org.mockito.Mockito.{mock, when => mockWhen}
 import org.scalatest.FunSuite
-import za.co.absa.enceladus.conformance.CmdConfig
-import za.co.absa.enceladus.conformance.interpreter.DynamicInterpreter
-import za.co.absa.enceladus.dao.EnceladusDAO
+import za.co.absa.enceladus.model.Dataset
+import za.co.absa.enceladus.model.conformanceRule.SparkSessionConfConformanceRule
 import za.co.absa.enceladus.samples.DeepArraySamples
 import za.co.absa.enceladus.utils.testUtils.SparkTestBase
 
-class SparkSessionRuleSuite extends FunSuite with SparkTestBase {
+class SparkSessionRuleSuite extends FunSuite with SparkTestBase with TestRuleBehaviors {
+
+  val sparkSessionRule = SparkSessionConfConformanceRule(order = 1, outputColumn = "TimeZone", controlCheckpoint = false, sparkConfKey = "spark.sql.session.timeZone")
+  val sparkSessionArrayRule = SparkSessionConfConformanceRule(order = 2, outputColumn = "items.SubSystem", controlCheckpoint = false, sparkConfKey = "spark.sql.session.mySubSystem")
+  val sparkSessionDeepArrayRule = SparkSessionConfConformanceRule(order = 3, outputColumn = "items.payments.SubSubSystem", controlCheckpoint = false, sparkConfKey = "spark.sql.session.mySubSubSystem")
+
+  val sparkSessionRulesList = List(sparkSessionRule, sparkSessionArrayRule, sparkSessionDeepArrayRule)
+
+  val sparkSessionOrdersDS = Dataset(name = "Orders Conformance", version = 1, hdfsPath = "src/test/testData/orders", hdfsPublishPath =
+    "testData/conformedOrders",
+    schemaName = "Orders", schemaVersion = 1,
+    conformance = sparkSessionRulesList)
+
+  val conformedSparkSessionOrdersJSON: String =
+    """{"id":1,"name":"First Order","items":[{"itemid":"ar229","qty":10,"price":5.1,"payments":[{"payid":"pid10","amount":51.0,"SubSubSystem":"FA2"}],"SubSystem":"FA1"},{"itemid":"2891k","qty":100,"price":1.1,"payments":[{"payid":"zk20","amount":100.0,"SubSubSystem":"FA2"}],"SubSystem":"FA1"},{"itemid":"31239","qty":2,"price":55.2,"payments":[],"SubSystem":"FA1"}],"errCol":[],"TimeZone":"GMT"}
+      |{"id":2,"name":"Second Order","items":[{"itemid":"AkuYdg","qty":100,"price":10.0,"payments":[{"payid":"d101","amount":10.0,"SubSubSystem":"FA2"},{"payid":"d102","amount":20.0,"SubSubSystem":"FA2"}],"SubSystem":"FA1"},{"itemid":"jUa1k0","qty":2,"price":55.2,"payments":[],"SubSystem":"FA1"}],"errCol":[],"TimeZone":"GMT"}
+      |{"id":3,"name":"Third Order","items":[{"itemid":"Gshj1","qty":10,"price":10000.0,"payments":[{"payid":"pid10","amount":2000.0,"SubSubSystem":"FA2"},{"payid":"pid10","amount":5000.0,"SubSubSystem":"FA2"}],"SubSystem":"FA1"},{"itemid":"Jdha2","qty":100,"price":45.0,"payments":[{"payid":"zk20","amount":150.0,"SubSubSystem":"FA2"},{"payid":"pid10","amount":2000.0,"SubSubSystem":"FA2"}],"SubSystem":"FA1"}],"errCol":[],"TimeZone":"GMT"}
+      |{"id":4,"name":"Fourth Order","items":[{"itemid":"dLda1","qty":10,"price":5.1,"payments":[{"payid":"pid10","amount":10.0,"SubSubSystem":"FA2"}],"SubSystem":"FA1"},{"itemid":"d2dhJ","qty":100,"price":1.1,"payments":[{"payid":"zk20","amount":15.0,"SubSubSystem":"FA2"}],"SubSystem":"FA1"},{"itemid":"Mska0","qty":2,"price":55.2,"payments":[],"SubSystem":"FA1"},{"itemid":"Gdal1","qty":20,"price":5.2,"payments":[],"SubSystem":"FA1"},{"itemid":"dakl1","qty":99,"price":1.2,"payments":[],"SubSystem":"FA1"}],"errCol":[],"TimeZone":"GMT"}
+      |{"id":5,"name":"Fifths order","items":[{"itemid":"hdUs1J","qty":50,"price":0.2,"payments":[{"payid":"pid10","amount":10.0,"SubSubSystem":"FA2"},{"payid":"pid10","amount":11.0,"SubSubSystem":"FA2"},{"payid":"pid10","amount":12.0,"SubSubSystem":"FA2"}],"SubSystem":"FA1"}],"errCol":[],"TimeZone":"GMT"}"""
+      .stripMargin.replace("\r\n", "\n")
+
 
   test("Spark Session conformance rule test") {
     val inputDf = spark.createDataFrame(DeepArraySamples.ordersData)
 
-    spark.conf.set("spark.sql.session.timeZone", "GMT")
     spark.conf.set("spark.sql.session.mySubSystem", "FA1")
     spark.conf.set("spark.sql.session.mySubSubSystem", "FA2")
 
-    //val v = spark.sessionState.conf.getConfString(rule.sparkConfKey)
-
-    inputDf.printSchema()
-    inputDf.show
-
-    implicit val dao: EnceladusDAO = mock(classOf[EnceladusDAO])
-    implicit val progArgs: CmdConfig = CmdConfig(reportDate = "2017-11-01")
-    implicit val enableCF: Boolean = false
-
-    mockWhen(dao.getDataset("Orders Conformance", 1)) thenReturn DeepArraySamples.sparkSessionOrdersDS
-
-    val mappingTablePattern = "{0}/{1}/{2}"
-
-    import spark.implicits._
-    val conformed = DynamicInterpreter.interpret(DeepArraySamples.sparkSessionOrdersDS, inputDf).cache
-
-    conformed.printSchema()
-    conformed.show
-
-    val conformedJSON = conformed.orderBy($"id").toJSON.collect().mkString("\n")
-
-    if (conformedJSON != DeepArraySamples.conformedSparkSessionOrdersJSON) {
-      println("EXPECTED:")
-      println(DeepArraySamples.conformedSparkSessionOrdersJSON)
-      println("ACTUAL:")
-      println(conformedJSON)
-      fail("Actual conformed dataset JSON does not match the expected JSON (see above).")
-    }
-
+    conformanceRuleShouldMatchExpected(inputDf, sparkSessionOrdersDS, conformedSparkSessionOrdersJSON)
   }
 
 }
